pyproject.toml
[tool.poetry]
name = "djchat"
version = "0.1.0"
description = ""
authors = ["SahilMehdiyev <branchsahil@gmail.com>"]
readme = "README.md"

packages = [
    { include = "server" }
]


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-dotenv = "^1.4.2"
python-dotenv = "^1.0.1"
django-jazzmin = "^3.0.1"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
